# apps/api/openapi.yaml
...
openapi: 3.0.0
paths:
...
  "/ml/analytics/transcription/costs":
    get:
      operationId: MLAnalyticsController_getTranscriptionCostSummary
      summary: Get transcription cost summary
      parameters:
      - name: hours
        required: false
        in: query
        description: 'Number of hours to look back (default: 24)'
        schema:
          type: number
      responses:
        '200':
          description: Cost summary retrieved successfully
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CostSummaryDto"
      tags:
      - ML Analytics
      security:
      - bearer: []
  "/ml/analytics/transcription/costs/detailed":
    get:
      operationId: MLAnalyticsController_getDetailedTranscriptionCosts
      summary: Get detailed transcription costs
      parameters:
      - name: hours
        required: false
        in: query
        description: 'Number of hours to look back (default: 24)'
        schema:
          type: number
      - name: limit
        required: false
        in: query
        description: 'Number of results to return (default: 100)'
        schema:
          type: number
      - name: offset
        required: false
        in: query
        description: 'Number of results to skip (default: 0)'
        schema:
          type: number
      responses:
        '200':
          description: Detailed costs retrieved successfully
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/DetailedCostsResponseDto"
      tags:
      - ML Analytics
      security:
      - bearer: []
  "/ml/analytics/tasks/summary":
    get:
      operationId: MLAnalyticsController_getMLTaskSummary
      summary: Get ML task summary across all types
      parameters:
      - name: hours
        required: false
        in: query
        description: 'Number of hours to look back (default: 24)'
        schema:
          type: number
      responses:
        '200':
          description: Task summary retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/MLResultSummaryDto"
      tags:
      - ML Analytics
      security:
      - bearer: []
  "/ml/analytics/budget/status":
    get:
      operationId: MLAnalyticsController_getBudgetStatus
      summary: Get current budget status
      parameters: []
      responses:
        '200':
          description: Budget status retrieved successfully
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BudgetStatusDto"
      tags:
      - ML Analytics
      security:
      - bearer: []
  "/ml/analytics/transcription/result/{shareId}":
    get:
      operationId: MLAnalyticsController_getTranscriptionResult
      summary: Get transcription result for a specific share
      parameters:
      - name: shareId
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Transcription result retrieved successfully
        '404':
          description: Transcription result not found
      tags:
      - ML Analytics
      security:
      - bearer: []
info:
  title: BookmarkAI API
  description: API for BookmarkAI service
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CostSummaryDto:
      type: object
      properties:
        periodHours:
          type: number
          description: Period in hours
        totalCostUsd:
          type: number
          description: Total cost in USD
        totalDurationSeconds:
          type: number
          description: Total audio duration in seconds
        totalDurationHours:
          type: number
          description: Total audio duration in hours
        transcriptionCount:
          type: number
          description: Number of transcriptions
        avgCostPerTranscription:
          type: number
          description: Average cost per transcription
        avgDurationSeconds:
          type: number
          description: Average duration in seconds
        costPerHour:
          type: number
          description: Cost per hour of audio
        breakdown:
          type: object
          description: Cost breakdown by backend
      required:
      - periodHours
      - totalCostUsd
      - totalDurationSeconds
      - totalDurationHours
      - transcriptionCount
      - avgCostPerTranscription
      - avgDurationSeconds
      - costPerHour
    TranscriptionCostDto:
      type: object
      properties:
        shareId:
          type: string
          description: Share ID
        audioDurationSeconds:
          type: number
          description: Audio duration in seconds
        billingUsd:
          type: number
          description: Cost in USD
        backend:
          type: string
          description: Backend used (api/local)
        createdAt:
          format: date-time
          type: string
          description: Creation timestamp
      required:
      - shareId
      - audioDurationSeconds
      - billingUsd
      - backend
      - createdAt
    DetailedCostsResponseDto:
      type: object
      properties:
        costs:
          description: List of costs
          type: array
          items:
            "$ref": "#/components/schemas/TranscriptionCostDto"
        total:
          type: number
          description: Total number of records
      required:
      - costs
      - total
    MLResultSummaryDto:
      type: object
      properties:
        taskType:
          type: string
          description: Task type
        count:
          type: number
          description: Number of tasks
        avgProcessingMs:
          type: number
          description: Average processing time in milliseconds
        totalProcessingMs:
          type: number
          description: Total processing time in milliseconds
        lastProcessedAt:
          format: date-time
          type: string
          description: Last processed timestamp
      required:
      - taskType
      - count
      - avgProcessingMs
      - totalProcessingMs
      - lastProcessedAt
    BudgetStatusDto:
      type: object
      properties:
        hourly:
          type: object
          description: Hourly budget status
        daily:
          type: object
          description: Daily budget status
      required:
      - hourly
      - daily